Module ocelot::oprf

Oblivious PRF traits + instantiations.

Modules

kkrt
Implementation of the batched, related-key oblivious pseudorandom function (BaRK-OPRF) protocol of Kolesnikov, Kumaresan, Rosulek, and Trieu (cf. https://eprint.iacr.org/2016/799, Figure 2).

kmprt
Implementation of the hash-based multi-use OPPRF of Kolesnikov, Matania, Pinkas, Rosulek, and Trieu (cf. https://eprint.iacr.org/2017/799).

Traits

ObliviousPrf
Trait containing the associated types used by an oblivious PRF.

Receiver
Trait for an oblivious PRF receiver.

Sender
Trait for an oblivious PRF sender.

Type Definitions

KkrtReceiver
KKRT oblivious PRF receiver using ALSZ OT extension with Chou-Orlandi as the base OT.

KkrtSender
KKRT oblivious PRF sender using ALSZ OT extension with Chou-Orlandi as the base OT.

KmprtReceiver
KMPRT hash-based OPPRF receiver, using KKRT as the underlying OPRF.

KmprtSender
KMPRT hash-based OPPRF sender, using KKRT as the underlying OPRF.