Module ocelot::ot

Oblivious transfer traits + instantiations.

This module provides traits for standard oblivious transfer (OT), correlated OT, and random OT, alongside implementations of the following OT protocols:

dummy: a dummy and completely insecure OT for testing purposes.

naor_pinkas: Naor-Pinkas semi-honest OT.

chou_orlandi: Chou-Orlandi malicious OT.

alsz: Asharov-Lindell-Schneider-Zohner semi-honest OT extension (+ correlated and random OT).

kos: Keller-Orsini-Scholl malicious OT extension (+ correlated and random OT).

Modules

alsz
Implementation of the Asharov-Lindell-Schneider-Zohner oblivious transfer extension protocol (cf. https://eprint.iacr.org/2016/602, Protocol 4).

chou_orlandi
Implementation of the Chou-Orlandi oblivious transfer protocol (cf. https://eprint.iacr.org/2015/267).

dummy
Implementation of an entirely insecure oblivious transfer protocol for testing purposes.

kos
Implementation of the Keller-Orsini-Scholl oblivious transfer extension protocol (cf. https://eprint.iacr.org/2015/546).

naor_pinkas
Implementation of the Naor-Pinkas oblivious transfer protocol (cf. https://dl.acm.org/citation.cfm?id=365502).

Traits

CorrelatedReceiver
Trait for one-out-of-two correlated oblivious transfer from the receiver's point-of-view.

CorrelatedSender
Trait for one-out-of-two correlated oblivious transfer from the sender's point-of-view.

RandomReceiver
Trait for one-out-of-two random oblivious transfer from the receiver's point-of-view.

RandomSender
Trait for one-out-of-two random oblivious transfer from the sender's point-of-view.

Receiver
Trait for one-out-of-two oblivious transfer from the receiver's point-of-view.

Sender
Trait for one-out-of-two oblivious transfer from the sender's point-of-view.

Type Definitions

AlszReceiver
Instantiation of the ALSZ OT extension receiver, using Chou-Orlandi as the base OT.

AlszSender
Instantiation of the ALSZ OT extension sender, using Chou-Orlandi as the base OT.

ChouOrlandiReceiver
Instantiation of the Chou-Orlandi OT receiver.

ChouOrlandiSender
Instantiation of the Chou-Orlandi OT sender.

DummyReceiver
Instantiation of the dummy OT receiver.

DummySender
Instantiation of the dummy OT sender.

KosReceiver
Instantiation of the KOS OT extension receiver, using Chou-Orlandi as the base OT.

KosSender
Instantiation of the KOS OT extension sender, using Chou-Orlandi as the base OT.

NaorPinkasReceiver
Instantiation of the Naor-Pinkas OT receiver.

NaorPinkasSender
Instantiation of the Naor-Pinkas OT sender.